Refactoring ini parsing code and separate it into a strategy pattern
