+++ src/LibreLancer.Data/Ini/IIniParser.cs
@@ -1,0 +1,14 @@
+﻿// MIT License - Copyright (c) Callum McGing
+// This file is subject to the terms and conditions defined in
+// LICENSE, which is part of this source code package
+
+using System.Collections.Generic;
+using System.IO;
+
+namespace LibreLancer.Ini
+{
+    public interface IIniParser
+    {
+        IEnumerable<Section> ParseIniFile(string path, Stream stream, bool preparse = true, bool allowmaps = false);
+    }
+}
